Add 'special keys' section of keyboard
This section includes a "Print" key and a "Back" Key. The print key
will only work if the user has JS Print Setup installed, and is
set to print silent (so no dialogs come up in a kiosk).
(https://addons.mozilla.org/en-US/firefox/addon/js-print-setup/)

The special key area is disabled by default. You can enable it
in the extensions prefs in about:config. The back button is defaulted
to show if the special key area is shown.